Code cleanup and move behaviors into altspace.utilities.behaviors
