+++ examples/steamvr/.eslintrc
@@ -1,0 +1,6 @@
+{
+    "parserOptions": {
+        "ecmaVersion": 6,
+        "sourceType": "module",
+    }
+}
