+++ package.json
@@ -18,6 +18,8 @@
     "browserify": "11.2.0",
     "conventional-github-releaser": "0.5.1",
     "del": "2.2.0",
+    "eslint": "^2.6.0",
+    "eslint-plugin-html": "^1.4.0",
     "gulp": "3.9.0",
     "gulp-awspublish": "3.0.1",
     "gulp-babel": "5.2.1",
@@ -41,8 +43,6 @@
     "run-sequence": "1.1.5",
     "vinyl-buffer": "1.0.0",
     "vinyl-source-stream": "1.1.0",
-    "yargs": "3.27.0",
-    "eslint": "^2.6.0",
-    "eslint-config-airbnb": "^6.2.0"
+    "yargs": "3.27.0"
   }
 }
